Debugando Interrupção primeira epoch
